adding error handling

+++ src/pages/components/DeleteButton.js
@@ -1,28 +1,36 @@
 const clearBucket = async () => {
-  try {
-    // const response = await fetch('http://localhost:8000/clear-bucket', {
-    const response = await fetch(
-      'https://waconverter.us-west-1.elasticbeanstalk.com/clear-bucket',
-      {
+  var confirmed = window.confirm('Are you sure you want to erase?');
+  const fetchButton = document.getElementById('fetch-btn');
+
+  if (confirmed) {
+    try {
+      const response = await fetch('http://localhost:8000/clear-bucket', {
+        // const response = await fetch(
+        //   'https://waconverter.us-west-1.elasticbeanstalk.com/clear-bucket',
+        //   {
         method: 'DELETE',
+      });
+
+      if (!response.ok) {
+        throw new Error('Failed to clear bucket');
       }
-    );
-
-    if (!response.ok) {
-      throw new Error('Failed to clear bucket');
+      fetchButton.click();
+      console.log('bucket cleared successfully');
+    } catch (error) {
+      console.error('Error clearing bucket:', error);
     }
-
-    console.log('bucket cleared successfully');
-  } catch (error) {
-    console.error('Error clearing bucket:', error);
+  } else {
+    console.log('Erase cancelled!');
   }
 };
 
 const DeleteButton = () => {
   return (
-    <div className="border-b-4">
-      <button onClick={clearBucket}>Clear conversions</button>
-    </div>
+    <>
+      <button id="deleteFiles-btn" onClick={clearBucket}>
+        Clear conversions
+      </button>
+    </>
   );
 };
